+++ Tests/PowerSyncTests/Kotlin/KotlinPowerSyncDatabaseImplTests.swift
@@ -26,7 +26,22 @@ final class KotlinPowerSyncDatabaseImplTests: XCTestCase {
         database = nil
         try await super.tearDown()
     }
-
+    
+    func testExecuteError() async throws {
+        do {
+            _ = try await database.execute(
+                sql: "INSERT INTO usersfail (id, name, email) VALUES (?, ?, ?)",
+                parameters: ["1", "Test User", "test@example.com"]
+            )
+            XCTFail("Expected an error to be thrown")
+        } catch {
+            XCTAssertEqual(error.localizedDescription, """
+error while compiling: INSERT INTO usersfail (id, name, email) VALUES (?, ?, ?)
+no such table: usersfail
+""")
+        }
+    }
+    
     func testInsertAndGet() async throws {
         _ = try await database.execute(
             sql: "INSERT INTO users (id, name, email) VALUES (?, ?, ?)",
@@ -47,6 +62,27 @@ final class KotlinPowerSyncDatabaseImplTests: XCTestCase {
         XCTAssertEqual(user.0, "1")
         XCTAssertEqual(user.1, "Test User")
         XCTAssertEqual(user.2, "test@example.com")
+    }
+    
+    func testGetError() async throws {
+        do {
+            let _ = try await database.get(
+                sql: "SELECT id, name, email FROM usersfail WHERE id = ?",
+                parameters: ["1"]
+            ) { cursor in
+                (
+                    try cursor.getString(name: "id"),
+                    try cursor.getString(name: "name"),
+                    try cursor.getString(name: "email")
+                )
+            }
+            XCTFail("Expected an error to be thrown")
+        } catch {
+            XCTAssertEqual(error.localizedDescription, """
+error while compiling: SELECT id, name, email FROM usersfail WHERE id = ?
+no such table: usersfail
+""")
+        }
     }
 
     func testGetOptional() async throws {
@@ -73,6 +109,27 @@ final class KotlinPowerSyncDatabaseImplTests: XCTestCase {
 
         XCTAssertEqual(existing, "Test User")
     }
+    
+    func testGetOptionalError() async throws {
+        do {
+            let _ = try await database.getOptional(
+                sql: "SELECT id, name, email FROM usersfail WHERE id = ?",
+                parameters: ["1"]
+            ) { cursor in
+                (
+                    try cursor.getString(name: "id"),
+                    try cursor.getString(name: "name"),
+                    try cursor.getString(name: "email")
+                )
+            }
+            XCTFail("Expected an error to be thrown")
+        } catch {
+            XCTAssertEqual(error.localizedDescription, """
+error while compiling: SELECT id, name, email FROM usersfail WHERE id = ?
+no such table: usersfail
+""")
+        }
+    }
 
     func testGetAll() async throws {
         _ = try await database.execute(
@@ -95,6 +152,27 @@ final class KotlinPowerSyncDatabaseImplTests: XCTestCase {
         XCTAssertEqual(users[0].1, "User 1")
         XCTAssertEqual(users[1].0, "2")
         XCTAssertEqual(users[1].1, "User 2")
+    }
+    
+    func testGetAllError() async throws {
+        do {
+            let _ = try await database.getAll(
+                sql: "SELECT id, name, email FROM usersfail WHERE id = ?",
+                parameters: ["1"]
+            ) { cursor in
+                (
+                    try cursor.getString(name: "id"),
+                    try cursor.getString(name: "name"),
+                    try cursor.getString(name: "email")
+                )
+            }
+            XCTFail("Expected an error to be thrown")
+        } catch {
+            XCTAssertEqual(error.localizedDescription, """
+error while compiling: SELECT id, name, email FROM usersfail WHERE id = ?
+no such table: usersfail
+""")
+        }
     }
 
     func testWatchTableChanges() async throws {
@@ -119,7 +197,7 @@ final class KotlinPowerSyncDatabaseImplTests: XCTestCase {
 
         let resultsStore = ResultsStore()
 
-        let stream = database.watch(
+        let stream = try database.watch(
             sql: "SELECT name FROM users ORDER BY id",
             parameters: nil
         ) { cursor in
@@ -127,7 +205,7 @@ final class KotlinPowerSyncDatabaseImplTests: XCTestCase {
         }
 
         let watchTask = Task {
-            for await names in stream {
+            for try await names in stream {
                 await resultsStore.append(names)
                 if await resultsStore.count() == 2 {
                     expectation.fulfill()
@@ -151,6 +229,29 @@ final class KotlinPowerSyncDatabaseImplTests: XCTestCase {
         let finalResults = await resultsStore.getResults()
         XCTAssertEqual(finalResults.count, 2)
         XCTAssertEqual(finalResults[1], ["User 1", "User 2"])
+    }
+    
+    func testWatchError() async throws {
+        do {
+            let stream = try database.watch(
+                sql: "SELECT name FROM usersfail ORDER BY id",
+                parameters: nil
+            ) { cursor in
+                cursor.getString(index: 0)!
+            }
+            
+            // Actually consume the stream to trigger the error
+            for try await _ in stream {
+                XCTFail("Should not receive any values")
+            }
+            
+            XCTFail("Expected an error to be thrown")
+        } catch {
+        XCTAssertEqual(error.localizedDescription, """
+error while compiling: EXPLAIN SELECT name FROM usersfail ORDER BY id
+no such table: usersfail
+""")
+        }
     }
 
     func testWriteTransaction() async throws {